+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.1.0
+
+- Enhancement: Retrieve coordinates of current active window [(#15)](https://github.com/nut-tree/libnut/issues/15)
+- Enhancement: Retrieve dimensions for a window specified via its window handle [(#17)](https://github.com/nut-tree/libnut/issues/17)
+- Enhancement: Separate folders for OS specific implementation [(#19)](https://github.com/nut-tree/libnut/issues/19)
+- Enhancement: Retrieve window name for a given window handle [(#22)](https://github.com/nut-tree/libnut/issues/22)
+
 ## 2.0.1
 
 - Bugfix: Fix hanging shift key after keyboard input on Windows [(#12)](https://github.com/nut-tree/libnut/issues/12)